refactored scheduler to their own individual components.. it's possible the subcomponents need additional refactoring
