+++ assets/fakeData.js
@@ -1,0 +1,63 @@
+export const fakeData = [
+    {
+      oc : 'open',
+      time: 1673506849,
+      id: '05299798-a44a-4fcd-a87e-fcc745236e68'
+    },
+    {
+      oc : 'closed',
+      time: 1673193829,
+      id:'3acdc454-8f96-4dc8-8ee9-f16afa322a69'
+    },
+    {
+      oc : 'closed',
+      time: 1673656849,
+      id:'4cc4ec13-1315-423f-9bfe-8151c56fc775'
+    },
+    {
+      oc : 'open',
+      time: 1673996849,
+      id: '1c73ef53-58ef-45b3-a167-44f9244fda9a'
+    },
+    {
+      oc : 'open',
+      time: 1673507955,
+      id: '1c73ef53-58ef-45b3-a167-44f9244fda9a'
+    },
+    {
+        oc : 'open',
+        time: 1773506849,
+        id: '05299798-a44a-5fcd-a87e-fcc745236e68'
+      },
+      {
+        oc : 'closed',
+        time: 1773193829,
+        id:'3acdc454-8f96-5dc8-8ee9-f16afa322a69'
+      },
+      {
+        oc : 'closed',
+        time: 1773656849,
+        id:'4cc4ec13-1315-523f-9bfe-8151c56fc775'
+      },
+      {
+        oc : 'open',
+        time: 1773996849,
+        id: '1c73ef53-58ef-55b3-a167-44f9244fda9a'
+      },
+      {
+        oc : 'open',
+        time: 1773507655,
+        id: '2c73ef53-58ef-55b3-a167-44f9244fda9a'
+      },
+      {
+        oc : 'open',
+        time: 1773996649,
+        id: '2c73ef53-58ef-55b3-a167-44f9244fda9a'
+      },
+      {
+        oc : 'open',
+        time: 1773507655,
+        id: '2c73ef53-58ef-55b3-a167-44f9244fda9a'
+      },
+  
+  ]
